requirements.txt
django==5.2.1
mysqlclient==2.2.7
djangorestframework==3.16.0
django-bootstrap5==25.1
python-dateutil==2.9.0.post0